# src/main/resources/static/swagger.yaml
openapi: 3.0.3
servers:
  - description: Localhost server
    url: http://localhost:8345
info:
  title: Country and Country Language API
  description: |
    This API provides information about countries and their languages.
    Additional Contacts:
      - Name: Another person
        Email: test@dddd.com
      - Name: Yet another person
        Email: [yetanother@dddd.com](mailto:yetanother@dddd.com)
  version: "1.0.0"
  contact:
    name: Marco T. Ávila Cerón
    email: marcotavilac@pm.me
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
  /country:
    get:
      tags:
        - countries
      summary: Fetch all countries
      description: Fetch all countries from the database
      operationId: fetchAllCountries
      responses:
        '200':
          description: A list of countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
    post:
      tags:
        - countries
      summary: Add a new country
      description: Add a new country to the database
      operationId: addCountry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                english_name:
                  type: string
      responses:
        '201':
          description: Country added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
...
components:
  schemas:
    CountryResponse:
      title: "CountryResponse"
      type: "object"
      properties:
        id:
          type: integer
          format: int32
        english_name:
          type: string
    Country:
      title: "Country"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int32"
        english_name:
          type: "string"
        languages:
          type: "array"
          items:
            $ref: "#/components/schemas/Language"
    Language:
      title: "Language"
      type: object
      required:
        - english_name
      properties:
        id:
          type: "integer"
          format: "int32"
        english_name:
          type: "string"